End care charge use case
