--- LBH.AdultSocialCare.Api/V1/UseCase/Common/Interfaces/ICancelCareChargeElementUseCase.cs
@@ -1,10 +1,0 @@
-using System;
-using System.Threading.Tasks;
-
-namespace LBH.AdultSocialCare.Api.V1.UseCase.Common.Interfaces
-{
-    public interface ICancelCareChargeElementUseCase
-    {
-        Task<bool> ExecuteAsync(Guid packageCareChargeId, Guid careElementId);
-    }
-}
